requirements.txt
# FPL Manager v3 - Production Dependencies

# Core Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Data Processing & Analysis
pandas==2.3.1
numpy==2.3.2

# Machine Learning & Advanced Analytics
scikit-learn==1.7.1
scipy==1.16.1
xgboost==1.7.6

# HTTP Requests & API Integration
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Database & Caching
sqlite3-python

# Visualization & Enhanced ML (Optional)
matplotlib==3.7.2
seaborn==0.12.2

# Development Tools
joblib==1.3.2
logging

# Weather API Integration
# (Using OpenWeatherMap via requests)

# News Integration 
# (Perplexity API via requests)

# Optional: Jupyter for development
# jupyter==1.0.0
# notebook==7.0.0